+++ src/csv/deser/niseci/mod.rs
@@ -28,6 +28,7 @@ use crate::csv::stanis::giorgio::csv_error_handler;
 #[allow(deprecated)]
 use crate::deser::validate_serialized_records;
 
+use crate::deser::limits::{ByteLimit, DefaultByteLimit};
 use crate::deser::{
     parse_serialized_records, RecordAnagraficaNISECI, RecordCampionamentoNISECI,
     RecordRiferimentoNISECI, TipoRecord,
@@ -171,7 +172,7 @@ where
     let config = CsvConfig::default()
         .with_delimiter(delimiter)
         .with_headers(has_headers);
-    check_riferimento_niseci_reader_conf::<R, T>(reader, config)
+    private_check_riferimento_niseci_reader::<R, DefaultByteLimit, T>(reader, config)
 }
 
 pub fn check_riferimento_niseci_reader_conf<R: Read, T>(
@@ -181,7 +182,17 @@ pub fn check_riferimento_niseci_reader_conf<R: Read, T>(
 where
     T: RecordRiferimentoNISECI + 'static,
 {
-    let normalizing_reader = NormalizerReader::new(reader);
+    private_check_riferimento_niseci_reader::<R, DefaultByteLimit, T>(reader, config)
+}
+
+fn private_check_riferimento_niseci_reader<R: Read, BL: ByteLimit, T>(
+    reader: R,
+    config: CsvConfig,
+) -> Result<Vec<T>, Vec<csv::Error>>
+where
+    T: RecordRiferimentoNISECI + 'static,
+{
+    let normalizing_reader = NormalizerReader::new(reader).take(BL::MAX_BYTES);
 
     let mut rdr = csv::ReaderBuilder::new()
         .delimiter(config.delimiter())
@@ -236,7 +247,7 @@ where
         return Err(err_vec);
     }
     let file = File::open(path).expect("Unable to open file");
-    check_riferimento_niseci_reader_conf(file, config)
+    private_check_riferimento_niseci_reader::<File, DefaultByteLimit, T>(file, config)
 }
 
 #[deprecated(
@@ -322,7 +333,7 @@ where
         _ => b',',
     };
 
-    check_campionamento_niseci_reader_conf::<R, T>(
+    private_check_campionamento_niseci_reader_conf::<R, DefaultByteLimit, T>(
         reader,
         CsvConfig::default()
             .with_delimiter(delimiter)
@@ -337,7 +348,17 @@ pub fn check_campionamento_niseci_reader_conf<R: Read, T>(
 where
     T: RecordCampionamentoNISECI + 'static,
 {
-    let normalizing_reader = NormalizerReader::new(reader);
+    private_check_campionamento_niseci_reader_conf::<R, DefaultByteLimit, T>(reader, config)
+}
+
+fn private_check_campionamento_niseci_reader_conf<R: Read, BL: ByteLimit, T>(
+    reader: R,
+    config: CsvConfig,
+) -> Result<Vec<T>, Vec<csv::Error>>
+where
+    T: RecordCampionamentoNISECI + 'static,
+{
+    let normalizing_reader = NormalizerReader::new(reader).take(BL::MAX_BYTES);
 
     let mut rdr = csv::ReaderBuilder::new()
         .delimiter(config.delimiter())
@@ -392,7 +413,7 @@ where
         return Err(err_vec);
     }
     let file = File::open(path).expect("Unable to open file");
-    check_campionamento_niseci_reader_conf(file, config)
+    private_check_campionamento_niseci_reader_conf::<File, DefaultByteLimit, T>(file, config)
 }
 
 #[deprecated(
@@ -521,7 +542,7 @@ where
         _ => b',',
     };
 
-    check_anagrafica_niseci_reader_conf::<R, T>(
+    private_check_anagrafica_niseci_reader_conf::<R, DefaultByteLimit, T>(
         reader,
         CsvConfig::default()
             .with_delimiter(delimiter)
@@ -536,7 +557,17 @@ pub fn check_anagrafica_niseci_reader_conf<R: Read, T>(
 where
     T: RecordAnagraficaNISECI + 'static,
 {
-    let normalizing_reader = NormalizerReader::new(reader);
+    private_check_anagrafica_niseci_reader_conf::<R, DefaultByteLimit, T>(reader, config)
+}
+
+fn private_check_anagrafica_niseci_reader_conf<R: Read, BL: ByteLimit, T>(
+    reader: R,
+    config: CsvConfig,
+) -> Result<Vec<T>, Vec<csv::Error>>
+where
+    T: RecordAnagraficaNISECI + 'static,
+{
+    let normalizing_reader = NormalizerReader::new(reader).take(BL::MAX_BYTES);
 
     let mut rdr = csv::ReaderBuilder::new()
         .delimiter(config.delimiter())
@@ -590,5 +621,5 @@ where
         return Err(err_vec);
     }
     let file = File::open(path).expect("Unable to open file");
-    check_anagrafica_niseci_reader_conf(file, config)
+    private_check_anagrafica_niseci_reader_conf::<File, DefaultByteLimit, T>(file, config)
 }