+++ CHANGELOG.md
@@ -22,6 +22,9 @@
   - Building with this feature turns on breaking changes
   - Usage is recommended but not default yet
   - The new APIs better reflect what `v0.2` will support
+- Add `deser::limits` module
+  - `deser::limits::ByteLimit` ensures bytes limit on reader methods
+  - Will allow external customisation in a future version
 
 ### Changed